Refine header
